pyproject.toml
[project]
name = "keboola-eshop"
version = "0.1.0"
description = "Local starter app for browsing Keboola Storage tables."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["keboola_eshop"]
